* backport/merge r1216:1291 from trunk (all changes except lib/sound)  * remove "Eidos Interactive" from the license statement in lib/sound because none of the code written there is owned/written by them
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/branches/sound@1292 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/keyedit.c
@@ -46,7 +46,7 @@
 #include "intimage.h"
 #include "lib/ivis_common/bitimage.h"
 #include "intdisplay.h"
-#include "audio_id.h"
+#include "lib/sound/audio_id.h"
 #include "lib/ivis_common/pieblitfunc.h"
 #include "multiint.h"
 
@@ -528,7 +528,7 @@ BOOL saveKeyMap(void)
 #define WRITE(var, size)                                               \
 	if (PHYSFS_write(pfile, var, 1, size) != size) {                     \
 		debug(LOG_ERROR, "saveKeyMap: could not write to %s %d bytes: %s", \
-		      KeyMapPath, size, PHYSFS_getLastError());                    \
+		      KeyMapPath, (int)size, PHYSFS_getLastError());                    \
 		assert(FALSE);                                                     \
 		return FALSE;                                                      \
 	}
